vm.py: added kwargs to ctor and json property generic
We know we at least want to add key_file_path or pem_key_string to vm
fields, but there are already so many params in the constructor we might
as well allow anything and just save the param to the vms attributes.

In addition, the vm.json property was hardcoded sending only some of the
vms fields. So instead of needing to add fields manually, better just
return all object fields, as we dont know what we may want in the
future.