hwprovisioner/Dockerfiles: dont build in stages
Building in stages is preferable as it makes the container a marginal
amount smaller, but it squashes it down to 1 layer, and therefore each
time it needs to be pushed, the 1 existing (heavy) layer needs to be
pushed.

Since it only makes the container a relatively insignificant amount
lighter, better to only push the changed layers after a rebuild.