cloudvisor/cleaner.py: periodically cleans dangling security groups and vms
The security groups dangle because when destroying a vm we dont wait for
it to be fully terminated, and its impossible to delete a security group
with resources. So the cleaner periodically cleans dangling security
groups.

In addition, it cleans dangling vms.. Really this should be unnecessary,
as the vms have an expiry date and should be expired, but because there
could potentially be a bug with that functionality and aws instances can
be expensive if left dangling for days, we prefer to cover this base as
well.