cloudvisor/rest.py: get base_image by tag instead of ami directly
If the rest sets the base_image by default directly to an ami then
whenever we would like to update the ami we will need to redeploy the
cloudvisor.

The idea instead is for the pytest client to send a base_image_tag in
the provisioning request, and that way we will not need to redeploy the
the backend just update the client to request whichever
base_image_version we would like.