cloudvisor/ec2_wrapper.py: only delete dangling security groups older than 5 minutes
Deleting vms takes a long time, so when requesting to delete a VM we
dont wait for it to finish, rather we send the delete command and return
success.
Security groups cant be deleting if they have resources which rely on
them.
These 2 facts make it necessary to run a secondary thread which wakes up
every X time and deletes dangling security groups, whose vms were
successfully deleted.
The problem is that the security group is obv created before the VM, so
there can be a rare situation where a security group is created, and
then before the vm is created the secondary thread deletes the security
group, because it thinks no vms rely on it (this is bc the vm is still
in the process of being created). This causes the vm creation to fail
(and therefore the allocation to fail), even though it shouldn't.

The easiest solution (for now) is to ignore security groups which were
created less then 5 mins ago when cleaning SGs, but creation time isn't
saved in the SG object. So to work around that I added a creation tag to
the SG, so that we can keep track of when they were created and delete
accordingly.