ec2_wrapper.py: allocate_vm in 2 stages
Up till now, allocate_vm would create_instance and wait for it to be
running before returning.

This generally was ok, but under load, the instance.wait_for_running
could take ~20 seconds, causing the main thread to be stuck on that
function, without the ability to jump to other threads and return
responses as it isnt an async function.

The solution is to divide the allocate into 2 parts, create_instance and
await_running, and await them sequentially, enabling the loop to jump to
other threads while waiting.