cloudvisor: get_or_create_security_group with lock
Without the lock, when requesting more than 1 resource simultaniously,
there is a race condition when allocating a machine, because 2 calls to
get_or_create_sg will first try to find, and then create when it doesnt
exists, simultaniously, causing the second create to throw an error
secureity group already exists (because the previous resource already
created it.

The solution is to introduce a lock to the ec2_wrapper object, and hold
the lock when trying to get_or_create a sg.