hardvisor/remote_executor.py: python wrapper for remote ssh execution
Between allocating resources to different users, the resource manager
needs to do some type of "cleaning", to make sure the previous user
didn't leave a mess on the machine.

In other resource managers (cloud, vm), the machine was simply deleted
and created from fresh. In physical machines this is a bit more
difficult because formatting the machine will take too much time.

For hardware devices the "cleaning" will need to be done via remote
(ssh) shell commands. That is the purpose of this remote_executor - to
make running remote shell commands asyncly and simply.