hardvisor/periodically.py: periodicaly clean dangling machines
If running the hardvisor with a provisioner, it can happen that they
will get "out of sync" - ie the hardvisor thinks that a specific machine
is in use, but the provisioner has already asked to free it. This
shouldn't happen, but nonetheless it can happen due to a bug.

The solution is to have a periodical cleaner which makes sure all
resource manager allocations exist in the provisioner as well, and if
not free the resource.