Cloudvisor: Changed logic of check_fulfill method
Currently, check_fulfill returns 'all or nothing' which means
it returns 'Success' if it can theorethically fulfill all machine requests
or 'Failed' otherwise. in order to support provisioning machines from different
resource managers, this logic required modification.

This changes the response of cloudvisor to Success even if it can provision
only part of the allocation requests, and it adds to the response, the hosts
it can provision