Provisioner: Add Support for allocating machines from different resource managers
The new design flow works this way:

The provisioner sends theoretically_fullfil request to all resource managers.
Each resource manager which can provide even some of the request's machines
return 'Success' and the list of machines it can allocate.

Then the fulfiller aggregates the answers and sends for each potential
fulfiller request for allocation of the machines it can allocate.
(currently not in parallel)

For each allocated host the fullfiller updates in redis the endpoint of
the resource manager which provided the machine, so it can send
deallocation request of dangling machines to their creator.

Then, provisoned hardware details are sent back to the client.

see: https://anyvision.atlassian.net/wiki/spaces/CS/pages/3489169546/Automation+infra+provisioner+-+HLD
for more info.